Extract AbstractBeanPropertyDefinition and PropertyFilterDefinition
