+++ server/src/main/java/com/vaadin/data/BeanPropertySet.java
@@ -21,7 +21,6 @@ import java.io.IOException;
 import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -33,10 +32,9 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import com.vaadin.data.BeanPropertySet.NestedBeanPropertyDefinition;
 import com.vaadin.data.util.BeanUtil;
 import com.vaadin.server.Setter;
-import com.vaadin.shared.util.SharedUtil;
-import com.vaadin.util.ReflectTools;
 
 /**
  * A {@link PropertySet} that uses reflection to find bean properties.
@@ -106,58 +104,6 @@ public class BeanPropertySet<T> implements PropertySet<T> {
         }
     }
 
-    private abstract static class AbstractBeanPropertyDefinition<T, V>
-            implements PropertyDefinition<T, V> {
-        private final PropertyDescriptor descriptor;
-        private final BeanPropertySet<T> propertySet;
-        private final Class<?> propertyHolderType;
-
-        public AbstractBeanPropertyDefinition(BeanPropertySet<T> propertySet,
-                Class<?> propertyHolderType, PropertyDescriptor descriptor) {
-            this.propertySet = propertySet;
-            this.propertyHolderType = propertyHolderType;
-            this.descriptor = descriptor;
-
-            if (descriptor.getReadMethod() == null) {
-                throw new IllegalArgumentException(
-                        "Bean property has no accessible getter: "
-                                + propertySet.beanType + "."
-                                + descriptor.getName());
-            }
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public Class<V> getType() {
-            return (Class<V>) ReflectTools
-                    .convertPrimitiveType(descriptor.getPropertyType());
-        }
-
-        @Override
-        public String getName() {
-            return descriptor.getName();
-        }
-
-        @Override
-        public String getCaption() {
-            return SharedUtil.propertyIdToHumanFriendly(getName());
-        }
-
-        @Override
-        public BeanPropertySet<T> getPropertySet() {
-            return propertySet;
-        }
-
-        protected PropertyDescriptor getDescriptor() {
-            return descriptor;
-        }
-
-        @Override
-        public Class<?> getPropertyHolderType() {
-            return propertyHolderType;
-        }
-    }
-
     private static class BeanPropertyDefinition<T, V>
             extends AbstractBeanPropertyDefinition<T, V> {
 
@@ -222,63 +168,6 @@ public class BeanPropertySet<T> implements PropertySet<T> {
          * @since 8.2
          */
         protected static final int MAX_PROPERTY_NESTING_DEPTH = 10;
-
-        /**
-         * Class containing the constraints for filtering nested properties.
-         *
-         * @since 8.2
-         *
-         */
-        protected static class PropertyFilterDefinition
-                implements Serializable {
-            private int maxNestingDepth;
-            private List<String> ignorePackageNamesStartingWith;
-
-            /**
-             * Create a property filter with max nesting depth and package names
-             * to ignore.
-             *
-             * @param maxNestingDepth
-             *            The maximum amount of nesting levels for
-             *            sub-properties.
-             * @param ignorePackageNamesStartingWith
-             *            Ignore package names that start with this string, for
-             *            example "java.lang".
-             */
-            public PropertyFilterDefinition(int maxNestingDepth,
-                    List<String> ignorePackageNamesStartingWith) {
-                this.maxNestingDepth = maxNestingDepth;
-                this.ignorePackageNamesStartingWith = ignorePackageNamesStartingWith;
-            }
-
-            /**
-             * Returns the maximum amount of nesting levels for sub-properties.
-             *
-             * @return maximum nesting depth
-             */
-            public int getMaxNestingDepth() {
-                return maxNestingDepth;
-            }
-
-            /**
-             * Returns a list of package name prefixes to ignore.
-             *
-             * @return list of strings that
-             */
-            public List<String> getIgnorePackageNamesStartingWith() {
-                return ignorePackageNamesStartingWith;
-            }
-
-            /**
-             * Get the default nested property filtering conditions.
-             *
-             * @return default property filter
-             */
-            public static PropertyFilterDefinition getDefaultFilter() {
-                return new PropertyFilterDefinition(MAX_PROPERTY_NESTING_DEPTH,
-                        Arrays.asList("java"));
-            }
-        }
 
         private final PropertyDefinition<T, ?> parent;
 
@@ -440,7 +329,7 @@ public class BeanPropertySet<T> implements PropertySet<T> {
 
     private static final ConcurrentMap<InstanceKey, BeanPropertySet<?>> INSTANCES = new ConcurrentHashMap<>();
 
-    private final Class<T> beanType;
+    final Class<T> beanType;
 
     private final Map<String, PropertyDefinition<T, ?>> definitions;
 
@@ -463,7 +352,7 @@ public class BeanPropertySet<T> implements PropertySet<T> {
     }
 
     private BeanPropertySet(Class<T> beanType, boolean checkNestedDefinitions,
-            NestedBeanPropertyDefinition.PropertyFilterDefinition propertyFilterDefinition) {
+            PropertyFilterDefinition propertyFilterDefinition) {
         this(beanType);
         if (checkNestedDefinitions) {
             Objects.requireNonNull(propertyFilterDefinition,
@@ -474,7 +363,7 @@ public class BeanPropertySet<T> implements PropertySet<T> {
 
     private void findNestedDefinitions(
             Map<String, PropertyDefinition<T, ?>> parentDefinitions, int depth,
-            NestedBeanPropertyDefinition.PropertyFilterDefinition filterCallback) {
+            PropertyFilterDefinition filterCallback) {
         if (depth >= filterCallback.getMaxNestingDepth()) {
             return;
         }
@@ -554,7 +443,7 @@ public class BeanPropertySet<T> implements PropertySet<T> {
     @SuppressWarnings("unchecked")
     public static <T> PropertySet<T> get(Class<? extends T> beanType,
             boolean checkNestedDefinitions,
-            NestedBeanPropertyDefinition.PropertyFilterDefinition filterDefinition) {
+            PropertyFilterDefinition filterDefinition) {
         Objects.requireNonNull(beanType, "Bean type cannot be null");
         InstanceKey key = new InstanceKey(beanType, false,
                 filterDefinition.getMaxNestingDepth(),
